Bit of refactoring
- Move reusable bits into ftm core
- Include entity id in export
